alx_travel_app/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.37.23
botocore==1.37.23
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-storages==1.14.5
django-taggit==6.1.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
drf-yasg==1.21.10
Faker==37.4.2
idna==3.10
inflection==0.5.1
jmespath==1.0.1
kombu==5.5.4
mysql-connector-python==9.2.0
mysqlclient==2.2.7
packaging==25.0
parameterized==0.9.0
pillow==11.1.0
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rabbitmq==0.2.0
requests==2.32.4
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
